week03_fast_pipelines/homework/.history/task3/requirements_20251215000907.txt
torch
torchvision
einops
gdown
pandas
tqdm
Pillow

